[🎨design]: Message 스타일 코드 추가
