+++ src/components/Common/Message/Message.tsx
@@ -1,8 +1,9 @@
 import MoreVertRoundedIcon from '@mui/icons-material/MoreVertRounded';
-import styled, { css } from 'styled-components';
 
 import { Avatar } from '@/components';
 import { Size } from '@/types';
+
+import { MessageContainer, MessageSender } from './Message.style';
 
 interface MessageProps {
   padding?: string;
@@ -46,44 +47,3 @@ export default function Message({
     </>
   );
 }
-
-interface MessageContainerProps {
-  $padding: string;
-  $fontSize: string;
-  $avatarSize: Size;
-}
-
-interface MessageSenderProps {
-  $fontSize: string;
-}
-
-const MessageContainer = styled.div<MessageContainerProps>`
-  ${({ $padding, $fontSize, $avatarSize, theme }) => css`
-    padding: ${$padding};
-    border: 2px solid red;
-
-    & > p {
-      padding: ${$padding};
-      padding-left: ${`calc(${theme.size.icon[$avatarSize]} + ${$padding} + 2px)`};
-      font-size: ${$fontSize};
-      word-break: break-all;
-      white-space: pre-wrap;
-    }
-  `}
-`;
-const MessageSender = styled.div<MessageSenderProps>`
-  ${({ $fontSize }) => css`
-    display: flex;
-    align-items: center;
-    font-size: ${$fontSize};
-    border: 2px solid orange;
-
-    & > span {
-      margin-left: 8px;
-    }
-
-    & > :last-child {
-      margin-left: auto;
-    }
-  `}
-`;
